+++ pywibeee/main.py
@@ -3,18 +3,20 @@
 
 """Command line interface (CLI) for WiBeee (old Mirubee) meter."""
 
-from pywibeee import __version__
-
 import argparse
+import asyncio
+import datetime
+import json
 import socket
-from xml.etree import ElementTree
 import sys
 import time
-import xmltodict
-import json
-import asyncio
-import httpx
 from multiprocessing import Process, Queue
+from xml.etree import ElementTree
+
+import httpx
+import xmltodict
+
+from pywibeee import __version__
 
 
 class WiBeee():
@@ -29,28 +31,28 @@ class WiBeee():
         self.data = None
         self.cmdport = 550
         self.model = None
-        self.modelDescription = None
+        self.model_description = None
         self.version = None
-        self.deviceName = None
+        self.device_name = None
         self.outformat = 'json'
         self.timedata = None
-        self.actionName = None
-        self._useAsync = 'async' in ucm
+        self.action_name = None
+        self._use_async = 'async' in ucm
 
-    def getIp(self):
+    def get_ip(self):
         """Get own IP."""
         s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
         s.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
         try:
             s.connect(('<broadcast>', 0))
             ip = s.getsockname()[0]
-        except Exception:
+        except OSError:
             ip = '127.0.0.1'
         return ip
 
-    def getSubnet(self):
+    def get_subnet(self):
         """Get own subnet."""
-        own_ip = self.getIp()
+        own_ip = self.get_ip()
         ip_split = own_ip.split('.')
         subnet = ip_split[:-1]
         subnetstr = '.'.join(subnet)
@@ -66,7 +68,7 @@ class WiBeee():
         except socket.error:
             queue.put((False, address, port))
 
-    def checkSubnetOpenPort(self, subnet):
+    def check_subnet_open_port(self, subnet):
         """Check subnet for open port IPs."""
         q = Queue()
         processes = []
@@ -77,23 +79,23 @@ class WiBeee():
             processes.append(p)
             p.start()
         found_ips = []
-        for idx, p in enumerate(processes):
+        for p in processes:
             if p.exitcode is None:
                 p.terminate()
             else:
                 # If finished, check if the port was open
-                open_ip, address, port = q.get()
+                open_ip, address, _port = q.get()
                 if open_ip:
                     found_ips.append(address)
         #  Cleanup processes
-        for idx, p in enumerate(processes):
+        for p in processes:
             p.join()
         return found_ips
 
     def autodiscover(self):
         """Autodiscover WiBeee host."""
         found = False
-        found_ips = self.checkSubnetOpenPort(self.getSubnet())
+        found_ips = self.check_subnet_open_port(self.get_subnet())
         for host in found_ips:
             resource = f'http://{host}:{self._port}/en/login.html'
             result = self.callurl(resource)
@@ -103,20 +105,20 @@ class WiBeee():
                 break
         return found
 
-    async def autodiscoverAsync(self):
+    async def autodiscover_async(self):
         """Autodiscover async WiBeee host."""
         found = False
-        found_ips = self.checkSubnetOpenPort(self.getSubnet())
+        found_ips = self.check_subnet_open_port(self.get_subnet())
         for host in found_ips:
             resource = f'http://{host}:{self._port}/en/login.html'
-            result = await self.callurlAsync(resource)
+            result = await self.callurl_async(resource)
             if result and '<title>WiBeee</title>' in result:
                 self.host = host
                 found = True
                 break
         return found
 
-    def getStatus(self, printTxt=True):
+    def get_status(self, print_txt=True):
         """Provide status."""
         self.data = 'ERROR'
         if self.host is None:
@@ -128,25 +130,25 @@ class WiBeee():
             result = self.callurl(resource)
             if result:
                 self.data = result
-        if printTxt:
-            return self.outputStatus()
+        if print_txt:
+            return self.output_status()
 
-    async def getStatusAsync(self, printTxt=True):
+    async def get_status_async(self, print_txt=True):
         """Provide async status."""
         self.data = 'ERROR'
         if self.host is None:
-            foundhost = await self.autodiscoverAsync()
+            foundhost = await self.autodiscover_async()
         else:
             foundhost = True
         if foundhost:
             resource = f'http://{self.host}:{self._port}/en/status.xml'
-            result = await self.callurlAsync(resource)
+            result = await self.callurl_async(resource)
             if result:
                 self.data = result
-        if printTxt:
-            return self.outputStatus()
+        if print_txt:
+            return self.output_status()
 
-    def getDeviceName(self, printTxt=True):
+    def get_device_name(self, print_txt=True):
         """Provide device name."""
         result = 'ERROR'
         if self.host is None:
@@ -158,59 +160,58 @@ class WiBeee():
             result = self.callurl(resource)
             if result:
                 self.data = result
-                self.deviceName = self.getParameterXML('id')
-                if not self.deviceName:
-                    self.deviceName = 'ERROR'
-        if printTxt:
+                self.device_name = self.get_parameter_xml('id')
+                if not self.device_name:
+                    self.device_name = 'ERROR'
+        if print_txt:
             if self.outformat == 'json':
-                result = self.outputJsonParam('devicename', self.deviceName)
+                result = self.output_json_param('devicename', self.device_name)
             elif self.outformat == 'plain':
-                result = self.deviceName
+                result = self.device_name
             elif self.outformat == 'xml':
-                result = '<devicename>' + self.deviceName + '</devicename>'
+                result = '<devicename>' + self.device_name + '</devicename>'
             elif self.outformat == 'file':
                 if not self.version:
-                    self.getVersion(printTxt=False)
-                filename = self.deviceName + '_' + self.version + '.xml'
-                f = open(filename, 'w')
-                f.write(self.deviceName)
-                f.close()
+                    self.get_version(print_txt=False)
+                filename = self.device_name + '_' + self.version + '.xml'
+                with open(filename, 'w', encoding='utf-8') as f:
+                    f.write(self.device_name)
                 result = 'File saved!'
             return result
 
-    async def getDeviceNameAsync(self, printTxt=True):
+    async def get_device_name_async(self, print_txt=True):
         """Provide async device name."""
         result = 'ERROR'
         if self.host is None:
-            foundhost = await self.autodiscoverAsync()
+            foundhost = await self.autodiscover_async()
         else:
             foundhost = True
         if foundhost:
             resource = f'http://{self.host}:{self._port}/services/user/devices.xml'
-            result = await self.callurlAsync(resource)
+            result = await self.callurl_async(resource)
             if result:
                 self.data = result
-                self.deviceName = self.getParameterXML('id')
-                if not self.deviceName:
-                    self.deviceName = 'ERROR'
-        if printTxt:
+                self.device_name = self.get_parameter_xml('id')
+                if not self.device_name:
+                    self.device_name = 'ERROR'
+        if print_txt:
             if self.outformat == 'json':
-                result = self.outputJsonParam('devicename', self.deviceName)
+                result = self.output_json_param('devicename', self.device_name)
             elif self.outformat == 'plain':
-                result = self.deviceName
+                result = self.device_name
             elif self.outformat == 'xml':
-                result = '<devicename>' + self.deviceName + '</devicename>'
+                result = '<devicename>' + self.device_name + '</devicename>'
             elif self.outformat == 'file':
                 if not self.version:
-                    self.getVersion(printTxt=False)
-                filename = self.deviceName + '_' + self.version + '.xml'
-                f = open(filename, 'w')
-                f.write(self.deviceName)
-                f.close()
+                    self.get_version(print_txt=False)
+                filename = self.device_name + '_' + self.version + '.xml'
+                with open(filename, 'w', encoding='utf-8') as f:
+                    f.write(self.device_name)
                 result = 'File saved!'
             return result
 
     def get_or_create_eventloop(self):
+        """Get or create an asyncio event loop for the current thread."""
         try:
             return asyncio.get_event_loop()
         except RuntimeError as ex:
@@ -219,43 +220,43 @@ class WiBeee():
                 asyncio.set_event_loop(loop)
                 return asyncio.get_event_loop()
 
-    def asyncCall(self, calling, printTxt=True):
+    def async_call(self, calling, print_txt=True):
         """Provide async call."""
         r = None
         loop = asyncio.get_event_loop()
-        if calling == 'getStatus':
-            r = loop.run_until_complete(self.getStatusAsync(printTxt))
+        if calling == 'get_status':
+            r = loop.run_until_complete(self.get_status_async(print_txt))
         elif calling == 'autodiscover':
-            r = loop.run_until_complete(self.autodiscoverAsync())
-        elif calling == 'getInfo':
-            r = loop.run_until_complete(self.getInfoAsync())
-        elif calling == 'getModel':
-            r = loop.run_until_complete(self.getModelAsync(printTxt))
-        elif calling == 'getDeviceName':
-            r = loop.run_until_complete(self.getDeviceNameAsync(printTxt))
-        elif calling == 'rebootWeb':
-            r = loop.run_until_complete(self.rebootWebAsync())
-        elif calling == 'resetEnergy':
-            r = loop.run_until_complete(self.resetEnergyAsync())
-        elif calling == 'configureServer':
-            # printTxt is (server_ip, server_port) tuple in this case
-            if isinstance(printTxt, tuple):
-                server_ip, server_port = printTxt
+            r = loop.run_until_complete(self.autodiscover_async())
+        elif calling == 'get_info':
+            r = loop.run_until_complete(self.get_info_async())
+        elif calling == 'get_model':
+            r = loop.run_until_complete(self.get_model_async(print_txt))
+        elif calling == 'get_device_name':
+            r = loop.run_until_complete(self.get_device_name_async(print_txt))
+        elif calling == 'reboot_web':
+            r = loop.run_until_complete(self.reboot_web_async())
+        elif calling == 'reset_energy':
+            r = loop.run_until_complete(self.reset_energy_async())
+        elif calling == 'configure_server':
+            # print_txt is (server_ip, server_port) tuple in this case
+            if isinstance(print_txt, tuple):
+                server_ip, server_port = print_txt
                 r = loop.run_until_complete(
-                    self.configureServerAsync(server_ip, server_port))
+                    self.configure_server_async(server_ip, server_port))
             else:
                 r = None
         return r
 
-    def setPort(self, port):
+    def set_port(self, port):
         """Set port value."""
         self._port = int(port)
 
-    def setTimeout(self, timeout):
+    def set_timeout(self, timeout):
         """Set timeout value."""
         self._timeout = float(timeout)
 
-    def setOutputFormat(self, outformat):
+    def set_output_format(self, outformat):
         """Set output format."""
         self.outformat = outformat
 
@@ -272,11 +273,11 @@ class WiBeee():
             result = 'Error: ConnectionError'
         except httpx.HTTPError:
             result = 'Error: RequestException'
-        except Exception as e:
+        except OSError as e:
             result = 'Error: ' + str(e)
         return result
 
-    async def callurlAsync(self, url):
+    async def callurl_async(self, url):
         """Call URL function using httpx (asynchronous)."""
         result = None
         try:
@@ -289,11 +290,11 @@ class WiBeee():
             result = 'Error: ConnectionError'
         except httpx.HTTPError:
             result = 'Error: RequestException'
-        except Exception as e:
+        except OSError as e:
             result = 'Error: ' + str(e)
         return result
 
-    def getInfo(self):
+    def get_info(self):
         """Get all info."""
         result = 'ERROR'
         if self.host is None:
@@ -302,137 +303,135 @@ class WiBeee():
             foundhost = True
         if foundhost:
             if not self.model:
-                self.getModel(printTxt=False)
-                self.getModelDescription()
+                self.get_model(print_txt=False)
+                self.get_model_description()
             if not self.version:
-                self.getVersion(printTxt=False)
-            if not self.deviceName:
-                self.getDeviceName(printTxt=False)
-        result = self.outputJsonParam('model', self.model)
-        result = self.outputJsonParam('model_description',
-                                      self.modelDescription, result)
-        result = self.outputJsonParam('webversion', self.version, result)
-        result = self.outputJsonParam('host', self.host, result)
-        result = self.outputJsonParam('devicename', self.deviceName, result)
+                self.get_version(print_txt=False)
+            if not self.device_name:
+                self.get_device_name(print_txt=False)
+        result = self.output_json_param('model', self.model)
+        result = self.output_json_param('model_description',
+                                        self.model_description, result)
+        result = self.output_json_param('webversion', self.version, result)
+        result = self.output_json_param('host', self.host, result)
+        result = self.output_json_param('devicename', self.device_name, result)
         return result
 
-    async def getInfoAsync(self):
+    async def get_info_async(self):
         """Get async all info."""
         result = 'ERROR'
         if self.host is None:
-            foundhost = await self.autodiscoverAsync()
+            foundhost = await self.autodiscover_async()
         else:
             foundhost = True
         if foundhost:
             if not self.model:
-                await self.getModelAsync(printTxt=False)
-                self.getModelDescription()
+                await self.get_model_async(print_txt=False)
+                self.get_model_description()
             if not self.version:
-                self.getVersion(printTxt=False)
-            if not self.deviceName:
-                await self.getDeviceNameAsync(printTxt=False)
-        result = self.outputJsonParam('model', self.model)
-        result = self.outputJsonParam('model_description',
-                                      self.modelDescription, result)
-        result = self.outputJsonParam('webversion', self.version, result)
-        result = self.outputJsonParam('host', self.host, result)
-        result = self.outputJsonParam('devicename', self.deviceName, result)
+                self.get_version(print_txt=False)
+            if not self.device_name:
+                await self.get_device_name_async(print_txt=False)
+        result = self.output_json_param('model', self.model)
+        result = self.output_json_param('model_description',
+                                        self.model_description, result)
+        result = self.output_json_param('webversion', self.version, result)
+        result = self.output_json_param('host', self.host, result)
+        result = self.output_json_param('devicename', self.device_name, result)
         return result
 
-    def outputJsonParam(self, parameter, value, oldData=None):
+    def output_json_param(self, parameter, value, old_data=None):
         """Provide output JSON parameter and value."""
-        if not oldData:
+        if not old_data:
             root = xmltodict.parse('<response></response>')
         else:
-            root = json.loads(oldData)
+            root = json.loads(old_data)
         if root['response'] is None:
             root['response'] = {}
         root['response'][parameter] = value
         return json.dumps(root)
 
-    def outputStatus(self, param=None, value=None):
+    def output_status(self):
         """Provide output."""
         result = 'ERROR'
         if not self.data:
-            self.getStatus()
-        if self.outformat == 'xml' or self.outformat == 'plain':
+            self.get_status()
+        if self.outformat in ('xml', 'plain'):
             result = self.data
         elif self.outformat == 'json':
             try:
                 root = xmltodict.parse(self.data)
-            except Exception:
+            except (xmltodict.expat.ExpatError, ElementTree.ParseError):
                 root = xmltodict.parse('<response></response>')
                 root['response'] = {}
                 root['response']['model'] = 'ERROR'
                 root['response']['model_description'] = 'ERROR'
                 root['response']['webversion'] = 'ERROR'
-                self.timedata = self.getTimeData()
+                self.timedata = self.get_time_data()
                 root['response']['time'] = self.timedata
                 result = json.dumps(root)
                 return result
             if 'model' not in self.data:
-                if self._useAsync:
-                    r = self.asyncCall('getModel', False)
+                if self._use_async:
+                    self.async_call('get_model', False)
                 else:
-                    self.getModel(printTxt=False)
-                if not 'response' in root:
+                    self.get_model(print_txt=False)
+                if 'response' not in root:
                     root['response'] = {}
                 root['response']['model'] = self.model
             else:
                 self.model = root['response']['model']
             if 'model_description' not in self.data:
-                self.getModelDescription()
-                if not 'response' in root:
+                self.get_model_description()
+                if 'response' not in root:
                     root['response'] = {}
-                root['response']['model_description'] = self.modelDescription
+                root['response']['model_description'] = self.model_description
             if 'webversion' not in self.data:
-                self.getVersion(printTxt=False)
-                if not 'response' in root:
+                self.get_version(print_txt=False)
+                if 'response' not in root:
                     root['response'] = {}
                 root['response']['webversion'] = self.version
             if 'time' not in self.data:
-                self.timedata = self.getTimeData()
-                if not 'response' in root:
+                self.timedata = self.get_time_data()
+                if 'response' not in root:
                     root['response'] = {}
                 root['response']['time'] = self.timedata
             result = json.dumps(root)
         elif self.outformat == 'file':
             if not self.model:
-                if self._useAsync:
-                    r = self.asyncCall('getModel', False)
+                if self._use_async:
+                    self.async_call('get_model', False)
                 else:
-                    self.getModel(printTxt=False)
-                self.getModelDescription()
+                    self.get_model(print_txt=False)
+                self.get_model_description()
             if not self.version:
-                self.getVersion(printTxt=False)
+                self.get_version(print_txt=False)
             filename = 'status_' + self.model + '_' + self.version + '.xml'
-            f = open(filename, 'w')
-            f.write(self.data)
-            f.close()
+            with open(filename, 'w', encoding='utf-8') as f:
+                f.write(self.data)
             result = 'File saved!'
         return result
 
-    def outputAction(self, data):
+    def output_action(self, data):
         """Provide output for action."""
         result = 'ERROR'
         if data:
             result = data
         if self.outformat == 'plain':
-            result = 'action ' + self.actionName + ' ' + result
+            result = 'action ' + self.action_name + ' ' + result
         elif self.outformat == 'xml':
-            result = ('<action id="' + self.actionName + '">' + result
+            result = ('<action id="' + self.action_name + '">' + result
                       + '</action>')
         elif self.outformat == 'json':
-            result = self.outputJsonParam(self.actionName, result)
+            result = self.output_json_param(self.action_name, result)
         elif self.outformat == 'file':
-            filename = 'action_' + self.actionName + '.txt'
-            f = open(filename, 'w')
-            f.write(result)
-            f.close()
+            filename = 'action_' + self.action_name + '.txt'
+            with open(filename, 'w', encoding='utf-8') as f:
+                f.write(result)
             result = 'File saved!'
         return result
 
-    def getModel(self, printTxt=True):
+    def get_model(self, print_txt=True):
         """Provide model."""
         result = 'ERROR'
         if self.host is None:
@@ -440,67 +439,65 @@ class WiBeee():
         else:
             foundhost = True
         if foundhost:
-            self.model = self.getParameterXML('model')
+            self.model = self.get_parameter_xml('model')
             if not self.model:
                 self.model = 'ERROR'
                 if foundhost:
-                    self.model = self.getModelWeb()
-            if self.modelDescription is None:
-                self.getModelDescription()
-        if printTxt:
+                    self.model = self.get_model_web()
+            if self.model_description is None:
+                self.get_model_description()
+        if print_txt:
             if self.outformat == 'json':
-                result = self.outputJsonParam('model', self.model)
-                result = self.outputJsonParam('model_description',
-                                      self.modelDescription, result)
+                result = self.output_json_param('model', self.model)
+                result = self.output_json_param('model_description',
+                                        self.model_description, result)
             elif self.outformat == 'plain':
-                result = self.model + ' ' + self.modelDescription
+                result = self.model + ' ' + self.model_description
             elif self.outformat == 'xml':
                 result = '<model>' + self.model + '</model>'
             elif self.outformat == 'file':
                 if not self.version:
-                    self.getVersion(printTxt=False)
+                    self.get_version(print_txt=False)
                 filename = self.model + '_' + self.version + '.xml'
-                f = open(filename, 'w')
-                f.write(self.model)
-                f.close()
+                with open(filename, 'w', encoding='utf-8') as f:
+                    f.write(self.model)
                 result = 'File saved!'
             return result
 
-    async def getModelAsync(self, printTxt=True):
+    async def get_model_async(self, print_txt=True):
         """Provide async model."""
         result = 'ERROR'
         if self.host is None:
-            foundhost = await self.autodiscoverAsync()
+            foundhost = await self.autodiscover_async()
         else:
             foundhost = True
         if foundhost:
-            self.model = self.getParameterXML('model')
+            self.model = self.get_parameter_xml('model')
             if not self.model:
                 self.model = 'ERROR'
                 if foundhost:
-                    self.model = await self.getModelWebAsync()
-            if self.modelDescription is None:
-                self.getModelDescription()
-        if printTxt:
+                    self.model = await self.get_model_web_async()
+            if self.model_description is None:
+                self.get_model_description()
+        if print_txt:
             if self.outformat == 'json':
-                result = self.outputJsonParam('model', self.model)
-                result = self.outputJsonParam('model_description',
-                                      self.modelDescription, result)
+                result = self.output_json_param('model', self.model)
+                result = self.output_json_param('model_description',
+                                        self.model_description, result)
             elif self.outformat == 'plain':
-                result = self.model + ' ' + self.modelDescription
+                result = self.model + ' ' + self.model_description
             elif self.outformat == 'xml':
                 result = '<model>' + self.model + '</model>'
             elif self.outformat == 'file':
                 if not self.version:
-                    self.getVersion(printTxt=False)
+                    self.get_version(print_txt=False)
                 filename = self.model + '_' + self.version + '.xml'
-                f = open(filename, 'w')
-                f.write(self.model)
-                f.close()
+                with open(filename, 'w', encoding='utf-8') as f:
+                    f.write(self.model)
                 result = 'File saved!'
             return result
 
-    def getModelWeb(self):
+    def get_model_web(self):
         """Provide model from web."""
         model = 'ERROR'
         data = ''
@@ -519,16 +516,16 @@ class WiBeee():
                 model = data[start+len(searchmodeltxt):end]
         return model
 
-    async def getModelWebAsync(self):
+    async def get_model_web_async(self):
         """Provide async model from web."""
         model = 'ERROR'
         data = ''
         resource = (f'http://{self.host}:{self._port}/en/loginRedirect.html'
                     '?user=user&pwd=user')
-        result = await self.callurlAsync(resource)
+        result = await self.callurl_async(resource)
         if result and 'Error' not in result:
             resource = f'http://{self.host}:{self._port}/en/index.html'
-            result = await self.callurlAsync(resource)
+            result = await self.callurl_async(resource)
             if result:
                 data = str(result)
             searchmodeltxt = 'var model = "'
@@ -554,77 +551,75 @@ class WiBeee():
         'WBP': 'Wibeee SMART PLUG',
     }
 
-    def getModelDescription(self):
+    def get_model_description(self):
         """Provide model description from model."""
-        self.modelDescription = self.MODEL_DESCRIPTIONS.get(
+        self.model_description = self.MODEL_DESCRIPTIONS.get(
             self.model, 'Unknown'
         )
 
-    def getTimeData(self):
+    def get_time_data(self):
         """Provide now time."""
-        import datetime
         now = datetime.datetime.now()
         return round(now.timestamp())
 
-    def getVersion(self, printTxt=True):
+    def get_version(self, print_txt=True):
         """Provide version from cmd."""
         result = 'ERROR'
-        self.version = self.getParameterXML('webversion')
+        self.version = self.get_parameter_xml('webversion')
         if not self.version:
-            resultCmd = self.doCmd(b'\x0d')
-            if resultCmd:
-                if not('Error:' in str(resultCmd)):
-                    cleanversion = resultCmd[3:]
+            result_cmd = self.do_cmd(b'\x0d')
+            if result_cmd:
+                if 'Error:' not in str(result_cmd):
+                    cleanversion = result_cmd[3:]
                     self.version = cleanversion.decode()
                 else:
-                    self.version = resultCmd
-        if printTxt:
+                    self.version = result_cmd
+        if print_txt:
             if self.outformat == 'json':
-                result = self.outputJsonParam('webversion', self.version)
+                result = self.output_json_param('webversion', self.version)
             elif self.outformat == 'plain':
                 result = self.version
             elif self.outformat == 'xml':
                 result = '<webversion>' + self.version + '</webversion>'
             elif self.outformat == 'file':
                 if not self.model:
-                    if self._useAsync:
-                        r = self.asyncCall('getModel', False)
+                    if self._use_async:
+                        self.async_call('get_model', False)
                     else:
-                        self.getModel(printTxt=False)
-                    self.getModelDescription()
+                        self.get_model(print_txt=False)
+                    self.get_model_description()
                 filename = self.model + '_' + self.version + '.xml'
-                f = open(filename, 'w')
-                f.write(self.version)
-                f.close()
+                with open(filename, 'w', encoding='utf-8') as f:
+                    f.write(self.version)
                 result = 'File saved!'
             return result
 
-    def readBackupPosition(self):
+    def read_backup_position(self):
         """Provide backup position."""
         result = None
-        self.actionName = 'readBackupPosition'
-        resultCmd = self.doCmd(b'\x0f')
-        if resultCmd:
-            if not('Error:' in resultCmd):
-                result = resultCmd.decode()
+        self.action_name = 'readBackupPosition'
+        result_cmd = self.do_cmd(b'\x0f')
+        if result_cmd:
+            if 'Error:' not in result_cmd:
+                result = result_cmd.decode()
             else:
-                result = resultCmd
-        return self.outputAction(result)
+                result = result_cmd
+        return self.output_action(result)
 
     def reboot(self):
         """Provide reboot from command."""
         result = None
-        self.actionName = 'reboot'
-        resultCmd = self.doCmd(b'\x01')
-        if resultCmd:
-            if not('Error:' in resultCmd):
-                result = resultCmd
-        return self.outputAction(result)
+        self.action_name = 'reboot'
+        result_cmd = self.do_cmd(b'\x01')
+        if result_cmd:
+            if 'Error:' not in result_cmd:
+                result = result_cmd
+        return self.output_action(result)
 
-    def rebootWeb(self):
+    def reboot_web(self):
         """Provide reboot from web."""
         result = None
-        self.actionName = 'rebootWeb'
+        self.action_name = 'rebootWeb'
         if self.host is None:
             foundhost = self.autodiscover()
         else:
@@ -634,27 +629,27 @@ class WiBeee():
             result = self.callurl(resource)
             if result and len(result) == 0:
                 result = 'done'
-        return self.outputAction(result)
+        return self.output_action(result)
 
-    async def rebootWebAsync(self):
+    async def reboot_web_async(self):
         """Provide async reboot from web."""
         result = None
-        self.actionName = 'rebootWeb'
+        self.action_name = 'rebootWeb'
         if self.host is None:
-            foundhost = await self.autodiscoverAsync()
+            foundhost = await self.autodiscover_async()
         else:
             foundhost = True
         if foundhost:
             resource = f'http://{self.host}:{self._port}/config_value?reboot=1'
-            result = await self.callurlAsync(resource)
+            result = await self.callurl_async(resource)
             if result and len(result) == 0:
                 result = 'done'
-        return self.outputAction(result)
+        return self.output_action(result)
 
-    def resetEnergy(self):
+    def reset_energy(self):
         """Provide reset energy from web."""
         result = None
-        self.actionName = 'resetEnergy'
+        self.action_name = 'resetEnergy'
         if self.host is None:
             foundhost = self.autodiscover()
         else:
@@ -664,24 +659,24 @@ class WiBeee():
             result = self.callurl(resource)
             if result and len(result) == 0:
                 result = 'done'
-        return self.outputAction(result)
+        return self.output_action(result)
 
-    async def resetEnergyAsync(self):
+    async def reset_energy_async(self):
         """Provide async reset energy from web."""
         result = None
-        self.actionName = 'resetEnergy'
+        self.action_name = 'resetEnergy'
         if self.host is None:
-            foundhost = await self.autodiscoverAsync()
+            foundhost = await self.autodiscover_async()
         else:
             foundhost = True
         if foundhost:
             resource = f'http://{self.host}:{self._port}/resetEnergy?resetEn=1'
-            result = await self.callurlAsync(resource)
+            result = await self.callurl_async(resource)
             if result and len(result) == 0:
                 result = 'done'
-        return self.outputAction(result)
+        return self.output_action(result)
 
-    def configureServer(self, server_ip, server_port=8600):
+    def configure_server(self, server_ip, server_port=8600):
         """Configure the WiBeee to push data to a server.
 
         The WiBeee firmware expects the port in hexadecimal format.
@@ -697,7 +692,7 @@ class WiBeee():
             Result string indicating success or failure.
         """
         result = 'ERROR'
-        self.actionName = 'configureServer'
+        self.action_name = 'configureServer'
         if self.host is None:
             foundhost = self.autodiscover()
         else:
@@ -719,14 +714,14 @@ class WiBeee():
                 )
                 self.callurl(resource)
                 result = f'done (server={server_ip}:{server_port})'
-        return self.outputAction(result)
+        return self.output_action(result)
 
-    async def configureServerAsync(self, server_ip, server_port=8600):
+    async def configure_server_async(self, server_ip, server_port=8600):
         """Provide async configure server push."""
         result = None
-        self.actionName = 'configureServer'
+        self.action_name = 'configureServer'
         if self.host is None:
-            foundhost = await self.autodiscoverAsync()
+            foundhost = await self.autodiscover_async()
         else:
             foundhost = True
         if foundhost:
@@ -737,18 +732,18 @@ class WiBeee():
                 f'&URLServidor={server_ip}'
                 f'&portServidor={port_hex}'
             )
-            result = await self.callurlAsync(resource)
+            result = await self.callurl_async(resource)
             if result is not None and 'Error' not in str(result):
                 # Reset the device to apply changes
                 reset_resource = (
                     f'http://{self.host}:{self._port}'
                     '/config_value?reset=true'
                 )
-                await self.callurlAsync(reset_resource)
+                await self.callurl_async(reset_resource)
                 result = f'done (server={server_ip}:{server_port})'
-        return self.outputAction(result)
+        return self.output_action(result)
 
-    def doCmd(self, cmd):
+    def do_cmd(self, cmd):
         """Provide cmd interface to the meter."""
         result = None
         # Create a TCP/IP socket
@@ -801,42 +796,41 @@ class WiBeee():
         'thd_ar9_V': ['THD Voltage Harmonic 9', 'V', 'mdi:numeric-9'],
     }
 
-    def getSensors(self):
+    def get_sensors(self):
         """Provide sensors list from XML."""
         result = 'ERROR'
-        sensorTypes = {}
+        sensor_types = {}
         if not self.data:
-            self.getStatus(False)
+            self.get_status(False)
         datavalue = str(self.data)
         try:
             xml = ElementTree.fromstring(datavalue)
-        except Exception:
+        except ElementTree.ParseError:
             xml = []
             return result
         for item in xml:
             if item.tag[:4] == 'fase':
                 name = item.tag[6:]
-                if name not in sensorTypes and name in self.SENSOR_DEFINITIONS:
-                    sensorTypes[name] = self.SENSOR_DEFINITIONS[name]
+                if name not in sensor_types and name in self.SENSOR_DEFINITIONS:
+                    sensor_types[name] = self.SENSOR_DEFINITIONS[name]
         if self.outformat == 'plain':
-            result = sensorTypes
+            result = sensor_types
         elif self.outformat == 'json':
-            result = json.dumps(sensorTypes)
+            result = json.dumps(sensor_types)
         elif self.outformat == 'xml':
-            result = sensorTypes
+            result = sensor_types
         elif self.outformat == 'file':
             filename = 'sensorTypes.txt'
-            f = open(filename, 'w')
-            f.write(json.dumps(sensorTypes))
-            f.close()
+            with open(filename, 'w', encoding='utf-8') as f:
+                f.write(json.dumps(sensor_types))
             result = 'File saved!'
         return result
 
-    def getParameterXML(self, param):
+    def get_parameter_xml(self, param):
         """Provide parameter value from XML."""
         parametervalue = None
         if not self.data:
-            self.getStatus(False)
+            self.get_status(False)
         datavalue = str(self.data)
         if param in datavalue:
             xml = ElementTree.fromstring(datavalue)
@@ -919,36 +913,36 @@ def program(args, printdata=True):
     frm = args.output
     if frm:
         if isinstance(frm, list):
-            c.setOutputFormat(frm[0])
+            c.set_output_format(frm[0])
         else:
-            c.setOutputFormat(frm)
+            c.set_output_format(frm)
     if args.auto:
-        c.asyncCall('autodiscover')
+        c.async_call('autodiscover')
     if args.get:
         if args.get[0] == 'model':
-            result = c.asyncCall('getModel')
+            result = c.async_call('get_model')
         elif args.get[0] == 'version':
-            result = c.getVersion()
+            result = c.get_version()
         elif args.get[0] == 'status':
-            result = c.asyncCall('getStatus')
+            result = c.async_call('get_status')
         elif args.get[0] == 'info':
-            result = c.asyncCall('getInfo')
+            result = c.async_call('get_info')
         elif args.get[0] == 'sensors':
-            result = c.getSensors()
+            result = c.get_sensors()
         elif args.get[0] == 'devicename':
-            result = c.asyncCall('getDeviceName')
+            result = c.async_call('get_device_name')
     if args.action:
         if args.action[0] == 'reboot':
             result = c.reboot()
         elif args.action[0] == 'rebootweb':
-            result = c.asyncCall('rebootWeb')
+            result = c.async_call('reboot_web')
         elif args.action[0] == 'resetenergy':
-            result = c.asyncCall('resetEnergy')
+            result = c.async_call('reset_energy')
         elif args.action[0] == 'configureserver':
-            server_ip = args.serverip[0] if args.serverip else c.getIp()
+            server_ip = args.serverip[0] if args.serverip else c.get_ip()
             server_port = args.serverport[0] if args.serverport else 8600
-            result = c.asyncCall(
-                'configureServer', (server_ip, server_port))
+            result = c.async_call(
+                'configure_server', (server_ip, server_port))
     if printdata:
         print(result)
     else: